[IMP] Moved up export_case_json_for_report in the business layer
