+++ source/app/iris_engine/reporter/reporter.py
@@ -21,19 +21,19 @@
 import logging as log
 import os
 from datetime import datetime
-
-from app.datamgmt.reporter.report_db import export_case_json_for_report
-from app.iris_engine.utils.common import IrisJinjaEnv
 from docx_generator.docx_generator import DocxGenerator
 from docx_generator.exceptions import rendering_error
 from flask_login import current_user
 from sqlalchemy import desc
 
 from app import app
+from app.business.cases import export_case_json_for_report
+from app.business.cases import export_case_json
+
 from app.datamgmt.activities.activities_db import get_auto_activities
 from app.datamgmt.activities.activities_db import get_manual_activities
 from app.datamgmt.case.case_db import case_get_desc_crc
-from app.business.cases import export_case_json
+
 from app.models import AssetsType
 from app.models import CaseAssets
 from app.models import CaseEventsAssets
@@ -43,7 +43,9 @@ from app.models import CasesEvent
 from app.models import Ioc
 from app.models import IocAssetLink
 from app.models import IocLink
+
 from app.iris_engine.reporter.ImageHandler import ImageHandler
+from app.iris_engine.utils.common import IrisJinjaEnv
 
 LOG_FORMAT = '%(asctime)s :: %(levelname)s :: %(module)s :: %(funcName)s :: %(message)s'
 log.basicConfig(level=log.INFO, format=LOG_FORMAT)
